refactor: Option 컴포넌트 자잘한 부분 수정
